pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astar-island"
version = "0.1.0"
description = "NM i AI 2026 — Astar Island solver (viewport observation + probabilistic terrain prediction)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.0",
    "pydantic>=2.0",
    "requests>=2.31",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "lightgbm>=4.0",
    "catboost>=1.2",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "httpx>=0.27"]

[project.scripts]
astar-island = "astar_island.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
